split lineage recovery tests
